requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
coverage==5.5
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
diart==0.9.0
docopt==0.6.2
docutils==0.20.1
einops==0.7.0
eyed3==0.9.7
ffmpeg==1.4
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
hmmlearn==0.3.2
huggingface-hub==0.21.3
HyperPyYAML==1.2.2
idna==3.6
imbalanced-learn==0.12.0
imblearn==0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
nh3==0.2.15
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20231117
optuna==3.5.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.19.0
pooch==1.8.1
primePy==1.3
protobuf==4.25.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==15.0.0
PyAudio==0.2.14
pyAudioAnalysis==0.3.14
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
readme_renderer==43.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
Rx==3.2.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.28
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
twine==5.0.0
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
ushlex==0.99.1
validators==0.22.0
wavio==0.0.8
websocket-client==1.7.0
websocket-server==0.6.4
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
